Restore old Offer accounting object and move dto into dto directory
