+++ src/main/groovy/life/qbic/datamodel/accounting/Offer.groovy
@@ -1,7 +1,5 @@
 package life.qbic.datamodel.accounting
 
-import groovy.transform.EqualsAndHashCode
-import life.qbic.datamodel.dtos.business.Customer
 import life.qbic.datamodel.dtos.business.OfferId
 import life.qbic.datamodel.dtos.general.Person
 
@@ -10,10 +8,11 @@ import life.qbic.datamodel.dtos.general.Person
  *
  * An offer describes a legally binding service proposal with associated costs.
  *
- * @since: 1.9.0*
+ * @since: 1.9.0
+ * @author: Tobias Koch
  *
  */
-@EqualsAndHashCode
+@Deprecated
 class Offer {
 
     /**
@@ -29,7 +28,7 @@ class Offer {
     /**
      * The customer for which this offer was created
      */
-    final Customer customer
+    final Person customer
 
     /**
      * The QBiC project manager who was assigned to the project
@@ -61,67 +60,24 @@ class Offer {
      */
     final OfferId identifier
 
-    static class Builder {
-
-        Date modificationDate
-
-        Date expirationDate
-
-        Customer customer
-
-        Person projectManager
-
-        String projectTitle
-
-        String projectDescription
-
-        List<ProductItem> items
-
-        double totalPrice
-
-        OfferId identifier
-
-        Builder(Date modificationDate, Date expirationDate, Customer customer, Person projectManager, String projectDescription, String projectTitle, List<ProductItem> items, double totalPrice, OfferId identifier) {
-            this.modificationDate = modificationDate
-            this.expirationDate = expirationDate
-            this.customer = customer
-            this.projectManager = projectManager
-            this.projectDescription = projectDescription
-            this.projectTitle = projectTitle
-            this.items = items
-            this.totalPrice = totalPrice
-            this.identifier = identifier
-        }
-        //ToDo Determine if any properties should be able to be modified later or can't be set to Null
-        Builder identifier(OfferId identifier) {
-            this.identifier = identifier
-            return this
-        }
-
-        Offer build() {
-            return new Offer(this)
-        }
+    Offer(Date modificationDate, Date expirationDate, Person customer, Person projectManager, String projectTitle, String projectDescription, List<ProductItem> items, double totalPrice, OfferId identifier) {
+        this.modificationDate = modificationDate
+        this.expirationDate = expirationDate
+        this.customer = customer
+        this.projectManager = projectManager
+        this.projectTitle = projectTitle
+        this.projectDescription = projectDescription
+        this.items = items
+        this.totalPrice = totalPrice
+        this.identifier = identifier
     }
 
-    private Offer(Builder builder) {
-        this.modificationDate = builder.modificationDate
-        this.expirationDate = builder.expirationDate
-        this.customer = builder.customer
-        this.projectManager = builder.projectManager
-        this.projectDescription = builder.projectDescription
-        this.projectTitle = builder.projectTitle
-        this.items = builder.items
-        this.totalPrice = builder.totalPrice
-        this.identifier = builder.identifier
-    }
-
-    //ToDo If an Offer is truly immutable then these methods should be removed
     /**
      * Adds a new item to the items list of the offer
      *
      * @param item which should be added to current list of items
      */
-    void addItem(ProductItem item) {
+    public void addItem(ProductItem item){
         items.add(item)
     }
 
@@ -130,14 +86,7 @@ class Offer {
      *
      * @param item which should be removed from the current list of items
      */
-    void removeItem(ProductItem item) {
+    void removeItem(ProductItem item){
         items.remove(item)
     }
-
-    //ToDo I don't think this method is needed.
-    @Override
-    String toString() {
-        return "${modificationDate.toString()} ${expirationDate.toString()}${customer.toString()} ${projectManager.toString()} ${projectDescription} ${projectTitle} ${items.toString()} ${totalPrice.toString()} ${identifier.toString()}"
-    }
 }
-
